pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "widget-gen"
version = "0.1.0"
description = "Проект для генерации изображений токенов"
authors = ["Your Name <you@example.com>"]
packages = [{include = "WidgetGen", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
# Другие зависимости вашего проекта...
aiofiles = "^23.2.1"
aniso8601 = "^9.0.1"
async-exit-stack = "^1.0.1"
async-generator = "^1.10"
certifi = "^2023.7.22"
chardet = "^5.2.0"
click = "^8.1.7"
dnspython = "^2.4.2"
email-validator = "^2.1.0.post1"
fastapi = "^0.104.1"
graphene = "^3.3"
graphql-core = "^3.2.3"
graphql-relay = "^3.2.0"
h11 = "^0.14.0"
httptools = "^0.6.1"
httpx = "^0.25.1"
idna = "^3.4"
itsdangerous = "^2.1.2"
jinja2 = "^3.1.2"
markupsafe = "^2.1.3"
orjson = "^3.9.10"
promise = "^2.3"
pydantic = "^2.5.1"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
pyyaml = "^6.0.1"
requests = "^2.31.0"
rx = "^3.2.0"
typing-extensions = "^4.8.0"
ujson = "^5.8.0"
urllib3 = "^2.1.0"
uvicorn = "^0.24.0.post1"
uvloop = "^0.19.0"
watchgod = "^0.8.2"
websockets = "^12.0"
pillow = "^10.1.0"

[tool.poetry.dev-dependencies]
black = "^21.4b0"
isort = "^5.8.0"
# Другие зависимости для разработки...

[tool.black]
line-length = 88
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
skip = [
    '.eggs', '.git', '.hg', '.mypy_cache', '.tox',
    '.venv', '_build', 'buck-out', 'build', 'dist',
]
skip_gitignore = true
